core changes
