+++ Entraineur.h
@@ -13,78 +13,14 @@ public:
     Entraîneur(int numéro, std::string nom, std::string téléphone, std::string date_naissance, double prix_heure)
         : numéro(numéro), nom(nom), téléphone(téléphone), date_naissance(date_naissance), embauche(""), salaire(0.0), prix_heure(prix_heure) {}
 
-    void afficherInformations() const {
-        std::cout << "Numéro: " << numéro << std::endl;
-        std::cout << "Nom: " << nom << std::endl;
-        std::cout << "Téléphone: " << téléphone << std::endl;
-        std::cout << "Date de Naissance: " << date_naissance << std::endl;
+    void afficherInformations() const;
 
-        if (!embauche.empty()) {
-            std::cout << "Date d'embauche: " << embauche << std::endl;
-            std::cout << "Salaire: " << salaire << std::endl;
-        } else {
-            std::cout << "Prix de l'heure: " << prix_heure << std::endl;
-        }
-
-        std::cout << std::endl;
-    }
-
-    int getNuméro() const {
-        return numéro;
-    }
-    std::string getNom() const {
-        return nom;
-    }
-    bool estPermanent() const {
-        return !embauche.empty();
-    }
+    int getNuméro() const;
+    std::string getNom() const;
+    bool estPermanent() const;
 
 private:
-    int numéro;class Equipement {
-public:
-    Equipement(std::string reference, std::string libelle, std::string dateAchat, double prixAchat)
-        : reference(reference), libelle(libelle), dateAchat(dateAchat), prixAchat(prixAchat), disponible(true) {}
-
-    std::string getReference() const {
-        return reference;
-    }
-
-    std::string getLibelle() const {
-        return libelle;
-    }
-
-    std::string getDateAchat() const {
-        return dateAchat;
-    }
-
-    double getPrixAchat() const {
-        return prixAchat;
-    }
-
-    bool estDisponible() const {
-        return disponible;
-    }
-
-    void marquerCommeIndisponible() {
-        disponible = false;
-    }
-
-private:
-    std::string reference;
-    std::string libelle;
-    std::string dateAchat;
-    double prixAchat;
-    bool disponible;
-};
-
-// Fonction pour afficher le menu
-void afficherMenu() {
-    std::cout << "Menu :" << std::endl;
-    std::cout << "1. Ajouter un équipement" << std::endl;
-    std::cout << "2. Afficher tous les équipements" << std::endl;
-    std::cout << "3. Marquer un équipement comme indisponible" << std::endl;
-    std::cout << "4. Quitter" << std::endl;
-}
+    int numéro;
     std::string nom;
     std::string téléphone;
     std::string date_naissance;
@@ -94,11 +30,7 @@ void afficherMenu() {
 };
 
 // Fonction pour afficher le menu
-void afficherMenuEnt() {
-    std::cout << "1. Ajouter un entraîneur" << std::endl;
-    std::cout << "2. Afficher tous les entraîneurs" << std::endl;
-    std::cout << "3. Supprimer un entraîneur" << std::endl;
-    std::cout << "4. Quitter" << std::endl;
-}
+void afficherMenu();
+void afficherMenuEnt();
 
 #endif // !ENTRAINEUR_H
