+++ main.cpp
@@ -1,4 +1,4 @@
-# include "Gestion.h"
+#include "SalleDeSport.h"
 
 int main() {
     SalleDeSport salleDeSport;
